split controller code to master and node controllers
Motivation behind this change is:
- we can make node controller as self contained, i.e, each node can create and
  delete volumes independent of controller.
- clearly divides what parts are handled by master and node, say for example
  nothing related to DeviceManager should be in master code.
- this is an enabler change to implement fully topology aware provisioning by
  dropping workaround solution of creating pmem device on attachment.
