--- pkg/pmem-csi-driver/controllerserver.go
@@ -1,548 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package pmemcsidriver
-
-import (
-	"fmt"
-	"strconv"
-
-	"github.com/google/uuid"
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	"k8s.io/klog/glog"
-
-	"github.com/container-storage-interface/spec/lib/go/csi"
-
-	pmemcommon "github.com/intel/pmem-csi/pkg/pmem-common"
-	pmdmanager "github.com/intel/pmem-csi/pkg/pmem-device-manager"
-	"k8s.io/utils/keymutex"
-)
-
-//VolumeStatus type representation for volume status
-type VolumeStatus int
-
-const (
-	//Created volume created
-	Created VolumeStatus = iota + 1
-	//Attached volume attached on a node
-	Attached
-	//Unattached volume dettached on a node
-	Unattached
-	//Deleted volume deleted
-	Deleted
-)
-
-type pmemVolume struct {
-	ID     string
-	Name   string
-	Size   int64
-	Status VolumeStatus
-	NodeID string
-	Erase  bool
-	NsMode string
-}
-
-type controllerServer struct {
-	*DefaultControllerServer
-	mode              DriverMode
-	rs                *registryServer
-	dm                pmdmanager.PmemDeviceManager
-	pmemVolumes       map[string]pmemVolume //Controller: map of reqID:VolumeInfo
-	publishVolumeInfo map[string]string     //Node: map of reqID:VolumeName
-}
-
-var _ csi.ControllerServer = &controllerServer{}
-var volumeMutex = keymutex.NewHashed(-1)
-var _ PmemService = &controllerServer{}
-
-func NewControllerServer(driver *CSIDriver, mode DriverMode, rs *registryServer, dm pmdmanager.PmemDeviceManager) *controllerServer {
-	serverCaps := []csi.ControllerServiceCapability_RPC_Type{}
-	switch mode {
-	case Controller:
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_CREATE_DELETE_VOLUME)
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_PUBLISH_UNPUBLISH_VOLUME)
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_LIST_VOLUMES)
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_GET_CAPACITY)
-	case Node:
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_PUBLISH_UNPUBLISH_VOLUME)
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_GET_CAPACITY)
-	case Unified:
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_CREATE_DELETE_VOLUME)
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_LIST_VOLUMES)
-		serverCaps = append(serverCaps, csi.ControllerServiceCapability_RPC_GET_CAPACITY)
-	}
-	driver.AddControllerServiceCapabilities(serverCaps)
-
-	return &controllerServer{
-		DefaultControllerServer: NewDefaultControllerServer(driver),
-		mode:                    mode,
-		rs:                      rs,
-		dm:                      dm,
-		pmemVolumes:             map[string]pmemVolume{},
-		publishVolumeInfo:       map[string]string{},
-	}
-}
-
-func (cs *controllerServer) RegisterService(rpcServer *grpc.Server) {
-	csi.RegisterControllerServer(rpcServer, cs)
-}
-
-func (cs *controllerServer) GetVolumeByID(volumeID string) *pmemVolume {
-	if pmemVol, ok := cs.pmemVolumes[volumeID]; ok {
-		return &pmemVol
-	}
-	return nil
-}
-
-func (cs *controllerServer) GetVolumeByName(Name string) *pmemVolume {
-	for _, pmemVol := range cs.pmemVolumes {
-		if pmemVol.Name == Name {
-			return &pmemVol
-		}
-	}
-	return nil
-}
-
-func (cs *controllerServer) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest) (*csi.CreateVolumeResponse, error) {
-	var vol *pmemVolume
-	eraseafter := true
-	nsmode := "fsdax"
-	if err := cs.Driver.ValidateControllerServiceRequest(csi.ControllerServiceCapability_RPC_CREATE_DELETE_VOLUME); err != nil {
-		pmemcommon.Infof(3, ctx, "invalid create volume req: %v", req)
-		return nil, err
-	}
-
-	if req.GetVolumeCapabilities() == nil {
-		return nil, status.Error(codes.InvalidArgument, "Volume Capabilities missing in request")
-	}
-
-	if len(req.GetName()) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "Name missing in request")
-	}
-	// Process plugin-specific parameters given as key=value pairs
-	params := req.GetParameters()
-	if params != nil {
-		// We recognize eraseafter=false/true, defaulting to true
-		for key, val := range params {
-			glog.Infof("CreateVolume: parameter: [%v] [%v]", key, val)
-			if key == "eraseafter" {
-				if val == "true" {
-					eraseafter = true
-				} else if val == "false" {
-					eraseafter = false
-				}
-			} else if key == "nsmode" {
-				if val == "fsdax" || val == "sector" {
-					nsmode = val
-				}
-			}
-		}
-	} else {
-		/* set default parameters */
-		params = map[string]string{
-			"nsmode":     nsmode,
-			"eraseafter": "true",
-		}
-	}
-
-	asked := req.GetCapacityRange().GetRequiredBytes()
-	// Required==zero means unspecified by CSI spec, we create a small 4 Mbyte volume
-	// as lvcreate does not allow zero size (csi-sanity creates zero-sized volumes)
-	if asked <= 0 {
-		asked = 4 * 1024 * 1024
-	}
-	topology := []*csi.Topology{}
-	glog.Infof("CreateVolume: Name: %v req.Required: %v req.Limit: %v", req.Name, asked, req.GetCapacityRange().GetLimitBytes())
-	if vol = cs.GetVolumeByName(req.Name); vol != nil {
-		// Check if the size of existing volume can cover the new request
-		glog.Infof("CreateVolume: Vol %s exists, Size: %v", vol.Name, vol.Size)
-		if vol.Size < asked {
-			return nil, status.Error(codes.AlreadyExists, fmt.Sprintf("Volume with the same name: %s but with different size already exist", vol.Name))
-		}
-	} else {
-		id, _ := uuid.NewUUID() //nolint: gosec
-		volumeID := id.String()
-		if cs.mode == Unified || cs.mode == Node {
-			glog.Infof("CreateVolume: Special create volume in Unified mode")
-			if err := cs.dm.CreateDevice(volumeID, uint64(asked), nsmode); err != nil {
-				return nil, status.Errorf(codes.Internal, "CreateVolume: failed to create volume: %s", err.Error())
-			}
-			topology = append(topology, &csi.Topology{
-				Segments: map[string]string{
-					"kubernetes.io/hostname": cs.Driver.nodeID,
-				},
-			})
-		} else /*if cs.mode == Controller */ {
-			capReq := &csi.GetCapacityRequest{
-				Parameters:         params,
-				VolumeCapabilities: req.GetVolumeCapabilities(),
-			}
-			foundNodes := []string{}
-			for _, node := range cs.rs.nodeClients {
-				cap, err := cs.getNodeCapacity(ctx, node, capReq)
-				if err != nil {
-					glog.Warningf("Error while fetching '%s' node capacity: %s", node.NodeID, err.Error())
-					continue
-				}
-				if cap >= asked {
-					glog.Infof("node %s has requested capacity", node.NodeID)
-					foundNodes = append(foundNodes, node.NodeID)
-				}
-			}
-
-			if len(foundNodes) == 0 {
-				return nil, status.Error(codes.Unavailable, fmt.Sprintf("No node found with %v capacity", asked))
-			}
-
-			glog.Infof("Found nodes: %v", foundNodes)
-
-			for _, id := range foundNodes {
-				topology = append(topology, &csi.Topology{
-					Segments: map[string]string{
-						"kubernetes.io/hostname": id,
-					},
-				})
-			}
-		}
-
-		vol = &pmemVolume{
-			ID:     volumeID,
-			Name:   req.GetName(),
-			Size:   asked,
-			Status: Created,
-			Erase:  eraseafter,
-			NsMode: nsmode,
-		}
-		if cs.mode == Unified || cs.mode == Node {
-			vol.Status = Attached
-		}
-		cs.pmemVolumes[volumeID] = *vol
-		glog.Infof("CreateVolume: Record new volume as %v", *vol)
-	}
-
-	return &csi.CreateVolumeResponse{
-		Volume: &csi.Volume{
-			VolumeId:           vol.ID,
-			CapacityBytes:      vol.Size,
-			AccessibleTopology: topology,
-			VolumeContext: map[string]string{
-				"name":   req.GetName(),
-				"size":   strconv.FormatInt(asked, 10),
-				"nsmode": nsmode,
-			},
-		},
-	}, nil
-}
-
-func (cs *controllerServer) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest) (*csi.DeleteVolumeResponse, error) {
-
-	// Check arguments
-	if len(req.GetVolumeId()) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "Volume ID missing in request")
-	}
-
-	if err := cs.Driver.ValidateControllerServiceRequest(csi.ControllerServiceCapability_RPC_CREATE_DELETE_VOLUME); err != nil {
-		pmemcommon.Infof(3, ctx, "invalid delete volume req: %v", req)
-		return nil, err
-	}
-
-	// Serialize by VolumeId
-	volumeMutex.LockKey(req.VolumeId)
-	defer volumeMutex.UnlockKey(req.VolumeId)
-
-	pmemcommon.Infof(4, ctx, "DeleteVolume: volumeID: %v", req.GetVolumeId())
-	vol := cs.GetVolumeByID(req.GetVolumeId())
-	if vol == nil {
-		pmemcommon.Infof(3, ctx, "Volume %s not created by this controller", req.GetVolumeId())
-	} else {
-		if vol.Status != Unattached {
-			pmemcommon.Infof(3, ctx, "Volume %s is attached to %s but not detached", vol.Name, vol.NodeID)
-		}
-		if cs.mode == Unified || cs.mode == Node {
-			glog.Infof("DeleteVolume: Special Delete in Unified mode")
-			if vol, ok := cs.pmemVolumes[req.GetVolumeId()]; ok {
-				if err := cs.dm.DeleteDevice(req.VolumeId, vol.Erase); err != nil {
-					return nil, status.Errorf(codes.Internal, "Failed to delete volume: %s", err.Error())
-				}
-			}
-		}
-		delete(cs.pmemVolumes, vol.ID)
-		pmemcommon.Infof(4, ctx, "DeleteVolume: volume %s deleted", req.GetVolumeId())
-	}
-	return &csi.DeleteVolumeResponse{}, nil
-}
-
-func (cs *controllerServer) ValidateVolumeCapabilities(ctx context.Context, req *csi.ValidateVolumeCapabilitiesRequest) (*csi.ValidateVolumeCapabilitiesResponse, error) {
-
-	// Check arguments
-	if len(req.GetVolumeId()) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "Volume ID missing in request")
-	}
-
-	_, found := cs.pmemVolumes[req.VolumeId]
-	if !found {
-		return nil, status.Error(codes.NotFound, "No volume found with id %s"+req.VolumeId)
-	}
-
-	if req.GetVolumeCapabilities() == nil {
-		return nil, status.Error(codes.InvalidArgument, "Volume capabilities missing in request")
-	}
-
-	for _, cap := range req.VolumeCapabilities {
-		if cap.GetAccessMode().GetMode() != csi.VolumeCapability_AccessMode_SINGLE_NODE_WRITER {
-			return &csi.ValidateVolumeCapabilitiesResponse{
-				Confirmed: nil,
-				Message:   "Driver does not support '" + cap.AccessMode.Mode.String() + "' mode",
-			}, nil
-		}
-	}
-
-	/*
-	 * FIXME(avalluri): Need to validate other capabilities against the existing volume
-	 */
-	return &csi.ValidateVolumeCapabilitiesResponse{
-		Confirmed: &csi.ValidateVolumeCapabilitiesResponse_Confirmed{
-			VolumeCapabilities: req.VolumeCapabilities,
-			VolumeContext:      req.GetVolumeContext(),
-		},
-	}, nil
-}
-
-func (cs *controllerServer) ListVolumes(ctx context.Context, req *csi.ListVolumesRequest) (*csi.ListVolumesResponse, error) {
-	pmemcommon.Infof(3, ctx, "ListVolumes")
-	if err := cs.Driver.ValidateControllerServiceRequest(csi.ControllerServiceCapability_RPC_LIST_VOLUMES); err != nil {
-		pmemcommon.Infof(3, ctx, "invalid list volumes req: %v", req)
-		return nil, err
-	}
-	// List namespaces
-	var entries []*csi.ListVolumesResponse_Entry
-	for _, vol := range cs.pmemVolumes {
-		info := &csi.Volume{
-			VolumeId:      vol.ID,
-			CapacityBytes: int64(vol.Size),
-			VolumeContext: map[string]string{
-				"name": vol.Name,
-				"size": strconv.FormatInt(vol.Size, 10),
-			},
-		}
-		entries = append(entries, &csi.ListVolumesResponse_Entry{
-			Volume:               info,
-			XXX_NoUnkeyedLiteral: *new(struct{}),
-			XXX_unrecognized:     nil,
-			XXX_sizecache:        0,
-		})
-	}
-
-	response := &csi.ListVolumesResponse{
-		Entries:              entries,
-		NextToken:            "",
-		XXX_NoUnkeyedLiteral: *new(struct{}),
-	}
-	return response, nil
-}
-
-func (cs *controllerServer) ControllerPublishVolume(ctx context.Context, req *csi.ControllerPublishVolumeRequest) (*csi.ControllerPublishVolumeResponse, error) {
-	if req.GetVolumeId() == "" {
-		return nil, status.Error(codes.InvalidArgument, "ControllerPublishVolume Volume ID must be provided")
-	}
-
-	if req.GetNodeId() == "" {
-		return nil, status.Error(codes.InvalidArgument, "ControllerPublishVolume Node ID must be provided")
-	}
-
-	if req.GetVolumeCapability() == nil {
-		return nil, status.Error(codes.InvalidArgument, "ControllerPublishVolume Volume capability must be provided")
-	}
-
-	// Serialize by VolumeId
-	volumeMutex.LockKey(req.VolumeId)
-	defer volumeMutex.UnlockKey(req.VolumeId)
-
-	glog.Infof("ControllerPublishVolume: cs.Node: %s req.volume_id: %s, req.node_id: %s ", cs.Driver.nodeID, req.VolumeId, req.NodeId)
-
-	if cs.mode == Controller {
-		vol, ok := cs.pmemVolumes[req.VolumeId]
-		if !ok {
-			return nil, status.Error(codes.NotFound, fmt.Sprintf("No volume with id '%s' found", req.VolumeId))
-		}
-		node, err := cs.rs.GetNodeController(req.NodeId)
-		if err != nil {
-			return nil, status.Error(codes.NotFound, err.Error())
-		}
-
-		conn, err := cs.rs.ConnectToNodeController(node.NodeID, connectionTimeout)
-		if err != nil {
-			return nil, status.Error(codes.Internal, err.Error())
-		}
-		defer conn.Close()
-
-		csiClient := csi.NewControllerClient(conn)
-		glog.Infof("Initiating Publishing volume ....")
-
-		req.VolumeContext = map[string]string{
-			"name":   vol.Name,
-			"size":   strconv.FormatInt(vol.Size, 10),
-			"nsmode": vol.NsMode,
-		}
-		resp, err := csiClient.ControllerPublishVolume(ctx, req)
-		glog.Infof("Got response")
-
-		if err == nil {
-			vol.Status = Attached
-			vol.NodeID = req.NodeId
-		}
-		return resp, err
-	}
-
-	if req.GetNodeId() != cs.Driver.nodeID {
-		// if asked nodeID does not match ours, return NotFound error
-		return nil, status.Error(codes.NotFound, "Node not found")
-	}
-	var volumeName string
-	var volumeSize uint64
-	var nsmode string
-	if cs.mode == Node {
-		attrs := req.GetVolumeContext()
-		if attrs == nil {
-			return nil, status.Error(codes.InvalidArgument, "Volume attributes must be provided")
-		}
-		volumeName = attrs["name"]
-		volumeSize, _ = strconv.ParseUint(attrs["size"], 10, 64)
-		nsmode = attrs["nsmode"]
-	} else /* if cs.mode == Unified */ {
-		vol, ok := cs.pmemVolumes[req.VolumeId]
-		if !ok {
-			return nil, status.Error(codes.NotFound, fmt.Sprintf("No volume with id '%s' found", req.VolumeId))
-		}
-		volumeName = vol.Name
-		volumeSize = uint64(vol.Size)
-		nsmode = vol.NsMode
-	}
-	// Check have we already published this volume.
-	// We may get called with same VolumeId repeatedly in short time
-	vol, published := cs.publishVolumeInfo[req.VolumeId]
-	if published {
-		glog.Infof("ControllerPublishVolume: Name:%v Id:%v already published, skip creation", vol, req.VolumeId)
-	} else {
-		glog.Infof("ControllerPublishVolume: volumeName:%v volumeSize:%v nsmode:%v", volumeName, volumeSize, nsmode)
-		/* Node/Unified */
-		if err := cs.dm.CreateDevice(req.VolumeId, volumeSize, nsmode); err != nil {
-			return nil, status.Errorf(codes.Internal, "Failed to create volume: %s", err.Error())
-		}
-
-		cs.publishVolumeInfo[req.VolumeId] = volumeName
-	}
-
-	return &csi.ControllerPublishVolumeResponse{
-		PublishContext: map[string]string{
-			"name": volumeName,
-			"size": strconv.FormatUint(volumeSize, 10),
-		},
-	}, nil
-}
-
-func (cs *controllerServer) ControllerUnpublishVolume(ctx context.Context, req *csi.ControllerUnpublishVolumeRequest) (*csi.ControllerUnpublishVolumeResponse, error) {
-	if req.GetVolumeId() == "" {
-		return nil, status.Error(codes.InvalidArgument, "ControllerPublishVolume Volume ID must be provided")
-	}
-
-	// Serialize by VolumeId
-	volumeMutex.LockKey(req.VolumeId)
-	defer volumeMutex.UnlockKey(req.VolumeId)
-
-	glog.Infof("ControllerUnpublishVolume : volume_id: %s, node_id: %s ", req.VolumeId, req.NodeId)
-
-	if cs.mode == Controller {
-		conn, err := cs.rs.ConnectToNodeController(req.NodeId, connectionTimeout)
-		if err != nil {
-			return nil, status.Error(codes.Internal, err.Error())
-		}
-
-		csiClient := csi.NewControllerClient(conn)
-		resp, err := csiClient.ControllerUnpublishVolume(ctx, req)
-		if err == nil {
-			if vol, ok := cs.pmemVolumes[req.GetVolumeId()]; ok {
-				vol.Status = Unattached
-			}
-		}
-
-		return resp, err
-	}
-
-	/* Node */
-
-	// eraseafter was possibly given as plugin-specific option,
-	// it was stored in the pmemVolumes entry during Create
-	erase := true
-	if vol, ok := cs.pmemVolumes[req.GetVolumeId()]; ok {
-		glog.Infof("ControllerUnpublish: use stored EraseAfter: %v", vol.Erase)
-		erase = vol.Erase
-	}
-	if err := cs.dm.DeleteDevice(req.GetVolumeId(), erase); err != nil {
-		return nil, status.Errorf(codes.Internal, "Failed to delete volume: %s", err.Error())
-	}
-
-	delete(cs.publishVolumeInfo, req.GetVolumeId())
-
-	return &csi.ControllerUnpublishVolumeResponse{}, nil
-}
-
-func (cs *controllerServer) GetCapacity(ctx context.Context, req *csi.GetCapacityRequest) (*csi.GetCapacityResponse, error) {
-	var capacity int64
-
-	if cs.mode == Controller {
-		for _, node := range cs.rs.nodeClients {
-			cap, err := cs.getNodeCapacity(ctx, node, req)
-			if err != nil {
-				glog.Warningf("Error while fetching '%s' node capacity: %s", node.NodeID, err.Error())
-				continue
-			}
-			capacity += cap
-		}
-
-		return &csi.GetCapacityResponse{
-			AvailableCapacity: capacity,
-		}, nil
-	}
-
-	/* Unified/Node */
-	cap, err := cs.dm.GetCapacity()
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, err.Error())
-	}
-
-	if params := req.GetParameters(); params != nil {
-		if nsmode, ok := params["nsmode"]; ok {
-			if c, ok := cap[nsmode]; ok {
-				capacity = int64(c)
-			} else {
-				return nil, fmt.Errorf("Unknown namespace mode :%s", nsmode)
-			}
-		}
-	}
-
-	return &csi.GetCapacityResponse{
-		AvailableCapacity: capacity,
-	}, nil
-}
-
-func (cs *controllerServer) getNodeCapacity(ctx context.Context, node NodeInfo, req *csi.GetCapacityRequest) (int64, error) {
-	conn, err := cs.rs.ConnectToNodeController(node.NodeID, connectionTimeout)
-	if err != nil {
-		return 0, fmt.Errorf("failed to connect to node %s: %s", node.NodeID, err.Error())
-	}
-
-	csiClient := csi.NewControllerClient(conn)
-	resp, err := csiClient.GetCapacity(ctx, req)
-	if err != nil {
-		return 0, fmt.Errorf("Error while fetching '%s' node capacity: %s", node.NodeID, err.Error())
-	}
-
-	return resp.AvailableCapacity, nil
-}
